A bunch of build and devtool updates
