+++ src/spaczz/__init__.py
@@ -1,6 +1,7 @@
 """spaczz library for fuzzy matching and extended regex functionality with spaCy."""
 try:
-    from importlib.metadata import version, PackageNotFoundError  # type: ignore
+    from importlib.metadata import PackageNotFoundError  # type: ignore
+    from importlib.metadata import version
 except ImportError:  # pragma: no cover
     from importlib_metadata import version, PackageNotFoundError  # type: ignore
