+++ src/spaczz/matcher/_phrasematcher.py
@@ -21,7 +21,8 @@ import warnings
 from spacy.tokens import Doc
 from spacy.vocab import Vocab
 
-from ..exceptions import KwargsWarning, PipeDeprecation
+from ..exceptions import KwargsWarning
+from ..exceptions import PipeDeprecation
 from ..search import _PhraseSearcher
 from ..util import nest_defaultdict
